pyproject.toml
[project]
name = "cse222-tinytapeout"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cocotb>=2.0.1",
    "cocotbext-uart>=0.1.4",
    "pytest>=9.0.2",
]
